Add function to display data to the I2C-LCD.
Along the way, fix a bug when initializing the master I2C clock pin.

The result is not satisfactory since the LCD is mostly unreadable
with the periodically updated data.

The Logic probe sees the occasional extra clock pulse and interprets
the write commands as reads.  Maybe the PIC16F1619 is doing likewise
and getting scrambled data.

An analog recording of the SDA and SCL signals does not show the extra
clock pulses, however, the appears some cross-talk between the signals.