Have the LCD showing the data fairly reliably.
Every so often there is a glitch in the i2c communication to the LCD
and the LCD gets crappy data.  The logic probe recording of the i2c
lines show extra clock pulses which mess with the write packets.
Maybe this is caused by the dangly wires used in the prototype.

The temporary fix is to write the data to the LCD only every few
measurements and to completely clear the display every 30 seconds
or so.